+++ backend/src/data/entities/key-result.ts
