+++ backend/src/data/entities/user.ts
@@ -1,6 +1,6 @@
 import { Entity, Column, ManyToOne } from 'typeorm';
 import { AbstractEntity } from '~/data/abstract/abstract.entity';
-import { DomainLevel } from './domainLevel';
+import { DomainLevel } from './domain-level';
 import { Company } from './company';
 
 @Entity()
